refactor: split monolithic models.py into db/models package
