+++ db/models/__init__.py
@@ -1,0 +1,13 @@
+from db.models.base import Base
+from db.models.user import User, UserRole, RefreshToken
+from db.models.document import Document, AnalysisResult, ScoreSummary
+
+__all__ = [
+    "Base",
+    "User",
+    "UserRole",
+    "RefreshToken",
+    "Document",
+    "AnalysisResult",
+    "ScoreSummary",
+]
