+++ db/models/document.py
@@ -1,73 +1,11 @@
-"""
-db/models.py — SQLAlchemy ORM model definitions.
-
-These models are database-agnostic (SQLAlchemy core + PostgreSQL JSONB).
-They contain no business logic and no connection details.
-"""
-
-import enum
 from datetime import datetime
 
-from sqlalchemy import (
-    Boolean,
-    Column,
-    DateTime,
-    Enum,
-    Float,
-    ForeignKey,
-    Integer,
-    String,
-    Text,
-)
+from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String, Text
 from sqlalchemy.dialects.postgresql import JSONB
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
-Base = declarative_base()
+from db.models.base import Base
 
-
-# ---------------------------------------------------------------------------
-# Auth models
-# ---------------------------------------------------------------------------
-
-class UserRole(str, enum.Enum):
-    free  = "free"
-    paid  = "paid"
-    admin = "admin"
-
-
-class User(Base):
-    __tablename__ = "users"
-
-    id               = Column(Integer, primary_key=True, index=True)
-    email            = Column(String, unique=True, index=True, nullable=False)
-    hashed_password  = Column(String, nullable=False)
-    role             = Column(Enum(UserRole), default=UserRole.free, nullable=False)
-    is_active        = Column(Boolean, default=True, nullable=False)
-    created_at       = Column(DateTime, default=datetime.utcnow)
-
-    refresh_tokens = relationship(
-        "RefreshToken",
-        back_populates="user",
-        cascade="all, delete-orphan",
-    )
-
-
-class RefreshToken(Base):
-    __tablename__ = "refresh_tokens"
-
-    id         = Column(Integer, primary_key=True)
-    token      = Column(String, unique=True, index=True, nullable=False)
-    user_id    = Column(Integer, ForeignKey("users.id"), nullable=False)
-    expires_at = Column(DateTime, nullable=False)
-    revoked    = Column(Boolean, default=False, nullable=False)
-
-    user = relationship("User", back_populates="refresh_tokens")
-
-
-# ---------------------------------------------------------------------------
-# ESG analysis models
-# ---------------------------------------------------------------------------
 
 class Document(Base):
     __tablename__ = "documents"
